Handle both suspending/unsuspending ports and disabling/enabling ports. If there is an array of vlans sent in the patch data, then remove the port from those vlans. Alternatively, if there is no vlans attribute sent with the patch data, then enable/disable the port.
As of this commit, there are two uses for this endpoint.
1. From the AccessR web interface, when clicking on "enable" or "disable" on a
switch port the vlans attribute is not included in the patch data and the
port is either enabled or disabled depending on its current state.

2. From the AccessX web interface, when clicking on "suspend" or "unsuspend"
on a provisioning record the vlans attribute is included causing AccessR to add
or remove the port from those vlans depending on its current state. The list of
vlans is configured in the .env variable VLANS_TO_SUSPEND. It is a quoted,
comma separated list of VLAN IDs, like "4,200,1000".