Unsuspend now uses Accessr::TalksToHardware
Not only for handling the ssh connection, it also uses the CrustyOldVendorConfig class to parse the existing rate_limits for the port. The input and output rate limits are removed from the config immediately before the port is re-added to the given vlans. Immediately after the port is re-added to the given vlans, the rate limit settings are re-inserted.